Added lib/core folder. Added RobotState Moved TeamUtils into lib/core
