+++ Robot Code/src/main/java/frc/lib/core/RobotState.java
@@ -1,0 +1,5 @@
+package frc.lib.core;
+
+public enum RobotState {
+    Disabled, Autonomous, Teleop, Test, Simulation
+}
