+++ Robot Code/src/main/java/frc/robot/PidParameters.java
@@ -2,6 +2,7 @@ package frc.robot;
 
 import java.util.Objects;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
+import frc.lib.core.TeamUtils;
 
 public class PidParameters {
   public static double telemetryUpdateInterval_secs = 0.5;
