+++ Robot Code/src/main/java/frc/robot/TeamSparkMAX.java
@@ -1,6 +1,7 @@
 package frc.robot;
 
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
+import frc.lib.core.TeamUtils;
 
 import com.revrobotics.CANSparkMax;
